requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
crispy-bootstrap5==0.7
Django==4.1.2
django-crispy-forms==1.14.0
django-extensions==3.2.1
et-xmlfile==1.1.0
gunicorn==20.1.0
haversine==2.7.0
idna==3.4
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
